requirements.txt
opencv-python
numpy
Pillow
scipy
torch
matplotlib
customtkinter
tksheet